bond payment added

+++ lib/models/bond/vm/bond_sembol_vm.dart
@@ -28,7 +28,7 @@ class BondInfo {
   String? symbol;
   int? v;
   int? couponFrequency;
-  int? couponRate;
+  double? couponRate;
   String? description;
   int? faceValue;
   int? issuePrice;
